feat: move logic from persons component
