+++ src/app/persons/experiences/experiences.component.html
@@ -1,0 +1,6 @@
+<nz-layout>
+    <nz-content>
+        <h1>Work experience of {{editedPerson?.name}} {{editedPerson?.surname}}</h1>
+        <app-experience-form [person]="editedPerson"></app-experience-form>
+    </nz-content>
+</nz-layout>
