+++ src/app/persons/experiences/experiences.component.spec.ts
@@ -1,28 +1,28 @@
 import { ComponentFixture, TestBed,  } from '@angular/core/testing';
 
-import { PersonsComponent } from './persons.component';
 import { ApolloTestingController, ApolloTestingModule } from 'apollo-angular/testing';
 import { ActivatedRoute, Router } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
-import { NgZorroAntdModule } from '../ng-zorro-antd.module';
-import { ExperienceFormComponent } from './experience-form/experience-form.component';
 import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
 import { Overlay } from '@angular/cdk/overlay';
+import { NgZorroAntdModule } from 'src/app/ng-zorro-antd.module';
+import { ExperienceFormComponent } from './experience-form/experience-form.component';
+import { ExperiencesComponent } from './experiences.component';
 
-describe('PersonsComponent', () => {
-  let component: PersonsComponent;
+describe('ExperiencesComponent', () => {
+  let component: ExperiencesComponent;
   let experienceFormComponent: ExperienceFormComponent;
-  let fixture: ComponentFixture<PersonsComponent>;
+  let fixture: ComponentFixture<ExperiencesComponent>;
   let experienceFormFixture: ComponentFixture<ExperienceFormComponent>;
   let controller: ApolloTestingController;
   let router: Router;  
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ PersonsComponent, ExperienceFormComponent ],
+      declarations: [ ExperiencesComponent, ExperienceFormComponent ],
       imports: [ RouterTestingModule.withRoutes(
         [
-          {path: 'persons/:id/experiences', component: PersonsComponent}
+          {path: 'Experiences/:id/experiences', component: ExperiencesComponent}
         ], {
           
         }
@@ -36,7 +36,7 @@ describe('PersonsComponent', () => {
     .compileComponents();
     controller = TestBed.inject(ApolloTestingController);
     router = TestBed.inject(Router);
-    fixture = TestBed.createComponent(PersonsComponent);
+    fixture = TestBed.createComponent(ExperiencesComponent);
     experienceFormFixture = TestBed.createComponent(ExperienceFormComponent);
     router.initialNavigation();
     component = fixture.componentInstance;
