+++ src/app/persons/persons-routing.module.ts
@@ -1,11 +1,18 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { PersonsComponent } from './persons.component';
+import { ExperiencesComponent } from './experiences/experiences.component';
 
 const routes: Routes = [
   {
-    path: ':id/experiences',
+    path: '',
     component: PersonsComponent,
+    children:[
+      {
+        path: ":id/experiences",
+        component: ExperiencesComponent
+      }
+    ]
   },
   
 ];
