+++ src/app/persons/persons.component.html
@@ -1,6 +1,1 @@
-<nz-layout>
-    <nz-content>
-        <h1>Work experience of {{editedPerson?.name}} {{editedPerson?.surname}}</h1>
-        <app-experience-form [person]="editedPerson"></app-experience-form>
-    </nz-content>
-</nz-layout>
+<router-outlet></router-outlet>
