+++ src/app/persons/persons.module.ts
@@ -3,14 +3,15 @@ import { CommonModule } from '@angular/common';
 
 import { PersonsRoutingModule } from './persons-routing.module';
 import { PersonsComponent } from './persons.component';
-import { ExperienceFormComponent } from './experience-form/experience-form.component';
+import { ExperienceFormComponent } from './experiences/experience-form/experience-form.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgZorroAntdModule } from '../ng-zorro-antd.module';
+import { ExperiencesComponent } from './experiences/experiences.component';
 
 
 @NgModule({
   declarations: [
-    PersonsComponent, ExperienceFormComponent
+    PersonsComponent, ExperienceFormComponent, ExperiencesComponent
   ],
   imports: [
     CommonModule,
